Remove unuse code

+++ src/main/java/com/phase_one/pythagorean/services/PythagoreanService.java
@@ -32,7 +32,7 @@ public class PythagoreanService {
 
     private PythagoreanValuesResponse calculateLegs(long c, long value) {
         long left = 1L;
-        long right = (long)c-1;
+        long right = c-1;
 
         while(left <= right){
             long computed = (left * left) + (right * right);
